Extract engine frame export operation
export_frames is a compact frame-sequence operation with format normalization, directory creation, and frame collection. This moves it into engine_frames.py while re-exporting export_frames from mcp_video.engine for compatibility.

Constraint: Preserve jpg/mjpeg/png handling, output naming, FFmpeg invocation, and ImageSequenceResult shape.

Rejected: Move create_from_images with export_frames | image-to-video creation is a larger inverse workflow and should be reviewed separately.

Confidence: high

Scope-risk: narrow

Directive: Keep frame sequence operations small until the larger image-sequence module is ready.

Tested: ruff check mcp_video/engine.py mcp_video/engine_frames.py

Tested: ruff format --check mcp_video/engine.py mcp_video/engine_frames.py

Tested: /opt/homebrew/bin/python3 export_frames re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine_advanced.py -k 'export_frames' tests/test_server.py -k 'export_frames' tests/test_cli.py -k 'export_frames or export-frames' -q --tb=short

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
